refactor(core): use u16 for page_count
- using u16 for page count information is sufficient and removes
unnecessary casts on most places
- Pager uses u16

[no changelog]

+++ core/embed/rust/src/ui/layout/obj.rs
@@ -312,7 +312,7 @@ impl LayoutObjInner {
 
         // Update page count if it changed
         if let Some(count) = self.event_ctx.page_count() {
-            self.page_count = count as u16;
+            self.page_count = count;
         }
 
         Ok(msg.into())
